refactor: move some functions from installer to workspace aspect
